[REFACTOR] #38: 장비 생성 응답값 수정

+++ src/main/java/cc/team3/character/dto/CharacterResponse.java
@@ -28,7 +28,7 @@ public class CharacterResponse {
 
     public record EquipmentDTO(
             String bonusType,
-            Integer bonusValue,
+            Double bonusValue,
             List<Effect> effects
     ) {
     }
